# Conflicts:
#	src/main/java/com/bacefook/controller/CommentsController.java
#	src/main/java/com/bacefook/controller/PostsController.java
#	src/main/java/com/bacefook/controller/UserController.java
#	src/main/java/com/bacefook/dao/CommentDAO.java
#	src/main/java/com/bacefook/entity/City.java
#	src/main/java/com/bacefook/entity/Comment.java
#	src/main/java/com/bacefook/entity/Gender.java
#	src/main/java/com/bacefook/entity/Post.java
#	src/main/java/com/bacefook/entity/PostLike.java
#	src/main/java/com/bacefook/entity/Relation.java
#	src/main/java/com/bacefook/entity/User.java
#	src/main/java/com/bacefook/entity/UserInfo.java
#	src/main/java/com/bacefook/repository/CommentsRepository.java
#	src/main/java/com/bacefook/repository/PostLikesRepository.java
#	src/main/java/com/bacefook/repository/RelationsRepository.java
#	src/main/java/com/bacefook/repository/UsersInfoRepository.java
#	src/main/java/com/bacefook/service/CommentLikeService.java
#	src/main/java/com/bacefook/service/PostLikesService.java
#	src/main/java/com/bacefook/service/RelationService.java
#	src/main/java/com/bacefook/service/UserService.java
#	src/main/resources/application.properties
#	src/test/java/com/bacefook/serviceTests/RelationsTest.java

+++ src/main/java/com/bacefook/controller/SessionManager.java
@@ -30,12 +30,12 @@ public class SessionManager {
 		Integer userId = (Integer) session.getAttribute(LOGGED_STATUS);
 //		session.setAttribute(LOGGED_STATUS, null);
 		session.invalidate();
-		return "User " + userId.intValue() + " was logged out!";
+		return "User " + userId+ " was logged out!";
 	}
 
 	/**
-	 * @return Logged user's ID, or null if no user is logged in
-	 * @throws UnauthorizedException
+	 * return Logged user's ID, or null if no user is logged in
+	 * throws UnauthorizedException
 	 */
 	public static Integer getLoggedUser(HttpServletRequest request) throws UnauthorizedException {
 		HttpSession session = request.getSession();
